core: added `appearance` and `editor layout` sub-menus
The commit adds the registration of an `appearance` and `editor layout` submenu to the
`view` main-menu.
Names have also been localized.
The `appearance` menu is used to group menu items related to the
appearance of the workbench or layout. The `editor layout` menu groups
items related to splitting the workbench. Commands present in vscode's
appearance and editor layout menu that exist in the framework were moved to these new
submenus.